SearchForQuery/requirements.txt
Flask==2.0.2
flask-marshmallow==0.14.0
gensim==3.8.3
Inflector==3.0.1
pandas==1.1.4
pymongo==3.12.1
scikit-learn==0.22
sklearn==0.0
stanfordcorenlp==3.8.0.1
tensorflow==2.3.1
tensorflow-estimator==2.3.0
tensorflow-hub==0.10.0
wikipedia==1.4.0
configparser==5.0.2
dill